+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
 ## [Unreleased]
+### Changed
+- Bumping up client version to support new features
+- All `@NotNull` annotations replaced with `@Nonnull`
+- Parameters extraction logic moved into the client
 
 ## [5.0.0-BETA-15]
 ### Added